refactor: authorize delete user by admin

+++ TaskManagement/src/main/java/org/redmath/taskmanagement/controller/UserController.java
@@ -31,7 +31,6 @@ public class UserController {
 
 
     @GetMapping
-    @PreAuthorize("hasRole('ADMIN')")
     public List<Users> getAllUsers() {
         return userService.getAllUsers();
     }
@@ -51,8 +50,12 @@ public class UserController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    @PreAuthorize("hasRole('ADMIN')")
+
     public void deleteUser(@PathVariable Long id, @AuthenticationPrincipal Jwt jwt) throws AccessDeniedException {
+
+        if (!jwt.getClaimAsStringList("roles").contains("ROLE_ADMIN")) {
+            throw new AccessDeniedException("You do not have permission to delete users");
+        }
         userService.deleteUser(id);
     }
 }
